+++ NEWS.md
@@ -1,3 +1,19 @@
+# eatGADS 0.18.1
+* `equalGADS()` now ignores irrelevant format differences (`F8.0` vs `F8`)
+* fix `JSS` link in vignette
+
+
+# eatGADS 0.18.0
+
+* `applyChangeMeta()` and `changeValLabels()` now automatically assign `"valid"` to the `missings` columns for labeled values
+* bug fix `equalGADS()` (meta data differences solely due to different meta data sorting are now ignored) 
+* `equalGADS()` can now exclude certain meta data columns from the comparison
+* `checkFormat()` for checking and modifying`SPSS` format type compatibility with actual data
+* `check4SPSS()` for checking `SPSS` meta data conventions and requirements
+* `changeVarNames()` and `applyChangeMeta()` now enforce `SQLite3` naming conventions
+* bug fixes `write_spss2()`
+* bug fix `applyChangeMeta()`: now recognizes differences in `NAs` in old columns when checking the change table
+
 
 # eatGADS 0.17.0
 
@@ -19,7 +35,7 @@
 
 # Internal
 * better performance of `applyChangeMeta()` on value level
-* bug fix `write_save()` (all columns are forced to be `double` when imported instead of `integer`)
+* bug fix `import_spss()` (all columns are forced to be `double` when imported instead of `integer`)
 * bug fix `fac2dummies_complex()` (value labels of dummy variables are tagged as `valid`)
 * bug fixes `recodeGADS()`
 * bug fix `removeValLabels()` (when a single value label was removed)